Improve form preview components for sample app
- Change FieldLayout from regular function to extension function WithLayout on BasicFormField
- Move Android-specific preview files to commonMain with @Preview annotations
