+++ internal/playground/src/commonMain/kotlin/soil/playground/form/compose/InputField.kt
@@ -3,12 +3,19 @@ package soil.playground.form.compose
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.text.input.VisualTransformation
+import org.jetbrains.compose.ui.tooling.preview.Preview
+import soil.form.FieldValidator
 import soil.form.compose.FormField
 import soil.form.compose.hasError
+import soil.form.compose.tooling.PreviewField
+import soil.form.rule.notBlank
+import soil.playground.style.AppTheme
 
 @Composable
 fun FormField<String>.InputField(
@@ -43,4 +50,22 @@ fun FormField<String>.InputField(
         keyboardActions = keyboardActions,
         visualTransformation = visualTransformation
     )
+}
+
+@Preview
+@Composable
+private fun InputFieldPreview() {
+    AppTheme {
+        PreviewField(
+            initialValue = "",
+            validator = FieldValidator {
+                notBlank { "This field cannot be blank" }
+            },
+            render = { field ->
+                field.WithLayout {
+                    InputField(label = { Text("Name") })
+                }
+            }
+        )
+    }
 }
